refactor(web): AssigneeDropdownをコンポーネントとして切り出す
